llvm: separate terminator lowering
